backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
pdfplumber==0.11.4
openpyxl==3.1.5
pandas==2.2.3
google-generativeai==0.8.4
python-dotenv==1.0.1
pydantic==2.10.4
